+++ src/config/index.js
@@ -179,6 +179,11 @@ let config = {
     errorHandlers: [], // callables for unhandled errors
     styleClassNameConverters,
   },
+  experimental: {
+    addBlockButton: {
+      enabled: false,
+    },
+  },
   widgets: {
     ...widgetMapping,
     default: defaultWidget,
@@ -204,6 +209,7 @@ let config = {
 };
 
 ConfigRegistry.settings = config.settings;
+ConfigRegistry.experimental = config.experimental;
 ConfigRegistry.blocks = config.blocks;
 ConfigRegistry.views = config.views;
 ConfigRegistry.widgets = config.widgets;